Run the add-token-counts script as a Docusaurus plugin post build
